replaced edge_traits with extended directional tag structures
